refactor: decompose planGate into focused sub-modules
Split the 776-line server/src/lib/planGate.ts into a directory of
focused modules (plan, features, usage, credits, checkout, assertions)
while preserving the public import path '../lib/planGate.js' via a
1-line barrel shim — no consumer changes required. All 238 tests across
the 18 planGate-touching test files remain green.
