+++ server/src/lib/planGate/index.ts
@@ -1,0 +1,76 @@
+// Barrel re-export — preserves the public API of the original
+// `lib/planGate.ts` so all existing imports of `'../lib/planGate.js'`
+// continue to resolve unchanged.
+
+
+export {
+  assertBinCreationAllowed,
+  assertBinCreationAllowedTx,
+  assertLocationWritable,
+  assertPhotoStorageAllowedTx,
+  assertReorganizeBinLimit,
+  checkLocationWritable,
+  getEffectiveMemberRole,
+} from './assertions.js';
+export {
+  buildDowngradeFlowAction,
+  buildPortalAction,
+  buildPortalUrl,
+  buildUpgradeAction,
+  buildUpgradePlanAction,
+  buildUpgradePlanUrl,
+  buildUpgradeUrl,
+  type CheckoutAction,
+  generateDowngradeFlowAction,
+  generatePortalAction,
+  generatePortalUrl,
+  generateUpgradeAction,
+  generateUpgradePlanAction,
+  generateUpgradePlanUrl,
+  generateUpgradeUrl,
+  getManagerToken,
+  getSubscriptionSecretKey,
+  renderActionAsUrl,
+} from './checkout.js';
+export {
+  type AiCreditInfo,
+  type AiCreditResult,
+  checkAndIncrementAiCredits,
+  getAiCredits,
+  refundAiCredit,
+} from './credits.js';
+export {
+  getFeatureMap,
+  getUserFeatures,
+  getUserLimitOverrides,
+  type PlanFeatures,
+  type UserLimitOverrides,
+} from './features.js';
+export {
+  getUserPlanInfo,
+  hasAiAccess,
+  isPlanRestricted,
+  isPlusOrAbove,
+  isProUser,
+  isSelfHosted,
+  isSubscriptionActive,
+  Plan,
+  type PlanTier,
+  planLabel,
+  SubStatus,
+  type SubStatusType,
+  subStatusLabel,
+  type UserPlanInfo,
+  validatePlanTransition,
+} from './plan.js';
+export {
+  computeOverLimits,
+  getUserBinCount,
+  getUserOverLimits,
+  getUserUsage,
+  invalidateOverLimitCache,
+  type OverLimits,
+  SELF_HOSTED_USAGE_STUB,
+  SELF_HOSTED_USAGE_SUMMARY_STUB,
+  type UserUsage,
+} from './usage.js';
